azure/requirements.txt
pandas==2.0.3
numpy==1.24.4
pythainlp==4.0.2
scikit-learn==1.3.2
protobuf==3.19.6
tensorflow-cpu==2.10.1
flask==2.2.5
azure-functions
protobuf==3.19.6